requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
weaviate-client==3.25.3
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.10
llama-cpp-python==0.2.27
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
